test: one pair of Lovelace resource collections
The registration tests and the removal tests each carried their own storage-mode
and YAML-mode collection. One pair in `tests/lovelace_helpers.py` now, keeping
the fuller storage-mode class — create and update as well as delete — and the
public `stored` name the removal tests read the backing store through.

Refs #230
